--- a/package.json
+++ b/package.json
@@ -57,7 +57,7 @@
   "dependencies": {
     "array-union": "^1.0.1",
     "dir-glob": "^2.0.0",
-    "fast-glob": "^2.0.1",
+    "fast-glob": "^2.0.2",
     "glob": "^7.1.2",
     "pify": "^3.0.0"
   },