@pvdlg 2018-01-31
For further improvements we could drop the `generateGlobTasks` function and just pass the `patterns` (with inverted patterns from `.gitignore`) to `fast-glob`.
We would have some cases that would work differently though:
- With `['a.tmp', '!*.tmp']` `fast-glob` returns `[]`, while `globby` returns `['a.tmp']`.
- But with `['tmp/a.tmp', '!tmp']` both returns `['tmp/a.tmp']`.

It seems that `fast-glob` allow negative patterns to be override by positive ones only for directory (i.e. the negative patter has to be a directory). But it always allow positive ones to override negative ones.