@pvdlg 2018-02-01
After doing a bit more test, it turns the current solution doesn't work properly for negative patterns in `.gitignore`, as files matched by those patterns end up being included in the result even if they are not part of the patterns passed by the user.

I added a failing test to demonstrate that.

`.gitignore` negative have to used as an exception in the scope of the files ignored within the same `.gitgnore` file. Unfortunately I don't see how this can be achieved without doing 2 steps as currently (glob then filter what's defined in the `.gitignore`).

If someone has a an idea how to solve this, let me know! Otherwise we'll have to revert to the old method. That said the old method can still be largely improved by refactoring the code, as the problem was the `.gitignore` file were read multiple times (once per pattern passed by the user). So we could improve the performance, by reading the `.gitignore` file before building the filter, rather than having the filter re-read them on every filter iteration.