@marionebl 2018-02-01
@pvdlg Just a rough idea: What if we combine both approaches, like this.

1. Read and parse relevant ignore files
2. Create ignore globs when safe, e.g. ignores with exclusion would be omitted. This should be considered only as performance optimization
3. Create a `filter` function as before and apply to `fast-glob` results to ensure all filter criteria are met