@pvdlg 2018-02-02
That sounds a good idea. I'm going to try that.

For info I did some test with the current `master` using `fast-glob`. With the current implementation (on master) I don't see much differences between `node-glob` and `fast-glob` regarding the cases with a `.gitignore`. I have an improvement a 1 to 3 op/s.

Not directly related, but I noticed that in most performance issue reported for XO, the performance issue come from what appear to be a bug in `node-glob` that is not present in `fast-glob`: isaacs/node-glob#309

```js
glob.sync('**/*.js', {ignore: ['**/node_modules/**']})
// => Works as expected, content of node_modules is excluded

glob.sync('./**/*.js', {ignore: ['**/node_modules/**']})
// => Content of node_modules is NOT excluded
```
This problem doesn't happen with `fast-glob`. It seems a lot of folks pass to XO a pattern that starts with `./` and that create a huge performance problem as non of the XO default ignores are used.

If we switch to `fast-glob` (without modifying the `.gitignore` handling) we would improve the performances a lot on those cases. With the case described in sindresorhus/xo#234, after setting Eslint to sue the same rules and plugins as XO I get:
- ~16s with ESLint
- ~30s with `xo ./**/*.js` 
- ~19s with `xo **/*.js` 
- ~16s with `xo **/*.js` if I replace the XO default ignores by just `**/node_modules/**`

=> Long story short, keeping the code as is and just replacing `node-glob` by `fast-glob` should get XO on par or pretty close to ESLint.