@sindresorhus 2018-02-03
> We could map nodir to onlyFiles to avoid a breaking change. Not sure if it's useful though, as some node-glob options are not available yet in fast-glob so it would be a breaking anyway.

Nah, we shouldn't map it. We need to do a major bump for this regardless.

> => Long story short, keeping the code as is and just replacing node-glob by fast-glob should get XO on par or pretty close to ESLint. Maybe we can do that, so we can get a XO release out, and then we'll continue to work on improving globby performances further.

I would be ok with that.