@sindresorhus 2018-02-03
Awesome work on investigating this @pvdlg :)