@sindresorhus 2019-01-09
@pvdlg Is there something worth merging from this PR? I could at least see the additional benchmarks and tests being useful.