@sindresorhus 2019-06-12
Closing this, but would really appreciate the mentioned PR with test/bench improvements whenever/if you have time.